refactor(external-browser): separate concerns across auth flow, listener, and payload
