+++ src/auth/external_browser.rs
@@ -1,18 +1,15 @@
-use std::io::{self, Write};
 use std::time::Duration;
 
-use reqwest::Client;
-use reqwest::Url;
+use reqwest::{Client, Url};
 use serde::Deserialize;
 use serde_json::json;
-use tokio::time;
 
 use crate::auth::client::AUTH_REQUEST_TIMEOUT;
 use crate::external_browser_launcher::{BrowserLauncher, LaunchOutcome, SystemCommandRunner};
 use crate::external_browser_listener::{
-    CallbackPayload, ListenerConfig, RunningListener, spawn_listener,
+    CallbackWaitError, ListenerConfig, RunningListener, spawn_listener,
 };
-use crate::external_browser_payload::parse_token_and_consent_from_pairs;
+use crate::external_browser_payload::BrowserCallbackPayload;
 use crate::{
     BrowserLaunchMode, Error, ExternalBrowserConfig, Result, WithCallbackListenerConfig,
     WithoutCallbackListenerConfig,
@@ -20,15 +17,22 @@ use crate::{
 
 #[cfg(unix)]
 mod manual_input_unix;
+mod manual_redirect_input;
 
 const EXTERNAL_BROWSER_CALLBACK_WAIT_TIMEOUT: Duration = Duration::from_secs(60);
 
-pub struct ExternalBrowserResult {
-    pub token: String,
-    pub proof_key: Option<String>,
+pub(crate) struct ExternalBrowserResult {
+    pub(crate) token: String,
+    pub(crate) proof_key: Option<String>,
 }
 
-pub async fn run_external_browser_flow(
+impl ExternalBrowserResult {
+    pub(crate) fn new(token: String, proof_key: Option<String>) -> Self {
+        Self { token, proof_key }
+    }
+}
+
+pub(in crate::auth) async fn run_external_browser_flow(
     http: &Client,
     username: &str,
     account: &str,
@@ -66,37 +70,37 @@ async fn run_external_browser_flow_with_listener(
     base_url: &Url,
     config_with_listener: &WithCallbackListenerConfig,
 ) -> Result<ExternalBrowserResult> {
-    let listener_config = ListenerConfig {
-        application: Some(super::client::client_app_id().to_string()),
-        host: config_with_listener.callback_socket_addr(),
-        port: config_with_listener.callback_socket_port(),
-        protocol: "http".to_string(),
-    };
+    let listener_config = ListenerConfig::new(
+        Some(super::client::client_app_id().to_string()),
+        config_with_listener.callback_socket_addr(),
+        config_with_listener.callback_socket_port(),
+        "http".to_string(),
+    );
     let listener = spawn_listener(listener_config)
         .await
         .map_err(|e| Error::Communication(e.to_string()))?;
-    let redirect_port = listener.addr.port();
+    let redirect_port = listener.redirect_port();
 
-    let auth = match request_authenticator(http, username, account, base_url, redirect_port).await {
-        Ok(data) => data,
-        Err(err) => {
-            shutdown_listener(listener).await;
-            return Err(err);
-        }
-    };
+    let auth =
+        match request_external_browser_challenge(http, username, account, base_url, redirect_port)
+            .await
+        {
+            Ok(data) => data,
+            Err(err) => {
+                listener.shutdown().await;
+                return Err(err);
+            }
+        };
 
     let launch_mode = config_with_listener.browser_launch_mode();
     if let Err(err) = open_auth_page(&auth.sso_url, launch_mode) {
-        shutdown_listener(listener).await;
+        listener.shutdown().await;
         return Err(err);
     }
 
     let payload =
         resolve_payload_with_listener(listener, EXTERNAL_BROWSER_CALLBACK_WAIT_TIMEOUT).await?;
-    Ok(ExternalBrowserResult {
-        token: payload.token,
-        proof_key: auth.proof_key,
-    })
+    Ok(ExternalBrowserResult::new(payload.token, auth.proof_key))
 }
 
 async fn run_external_browser_flow_without_listener(
@@ -107,30 +111,31 @@ async fn run_external_browser_flow_without_listener(
     config_without_listener: &WithoutCallbackListenerConfig,
 ) -> Result<ExternalBrowserResult> {
     let redirect_port = config_without_listener.redirect_port().get();
-    let auth = request_authenticator(http, username, account, base_url, redirect_port).await?;
+    let auth = request_external_browser_challenge(http, username, account, base_url, redirect_port)
+        .await?;
     let launch_mode = config_without_listener.browser_launch_mode();
     open_auth_page(&auth.sso_url, launch_mode)?;
 
     eprintln!(
         "The browser may display a connection error page for localhost:{redirect_port}; this is expected in this mode."
     );
-    let payload = manual_token_flow().await?;
+    let payload = manual_redirect_input::manual_token_flow().await?;
     Ok(ExternalBrowserResult {
         token: payload.token,
         proof_key: auth.proof_key,
     })
 }
 
-async fn request_authenticator(
+async fn request_external_browser_challenge(
     http: &Client,
     username: &str,
     account: &str,
     base_url: &Url,
     redirect_port: u16,
-) -> Result<AuthenticatorData> {
+) -> Result<ExternalBrowserChallenge> {
     let url = base_url.join("session/authenticator-request")?;
 
-    let body = authenticator_request_body(username, account, redirect_port);
+    let body = challenge_request_body(username, account, redirect_port);
 
     let resp = http
         .post(url)
@@ -144,7 +149,7 @@ async fn request_authenticator(
         return Err(Error::Communication(format!("HTTP {status}: {text}")));
     }
 
-    let parsed: AuthenticatorResponse =
+    let parsed: ExternalBrowserChallengeResponse =
         serde_json::from_str(&text).map_err(|e| Error::Json(e, text.clone()))?;
     if !parsed.success {
         return Err(Error::Communication(parsed.message.unwrap_or_default()));
@@ -152,12 +157,12 @@ async fn request_authenticator(
 
     let data = parsed
         .data
-        .ok_or_else(|| Error::Communication("missing authenticator-response data".to_string()))?;
+        .ok_or_else(|| Error::Communication("missing challenge response data".to_string()))?;
     Ok(data)
 }
 
 #[derive(Debug, Deserialize)]
-struct AuthenticatorData {
+struct ExternalBrowserChallenge {
     #[serde(rename = "ssoUrl")]
     sso_url: String,
     #[serde(rename = "proofKey")]
@@ -165,17 +170,13 @@ struct AuthenticatorData {
 }
 
 #[derive(Debug, Deserialize)]
-struct AuthenticatorResponse {
-    data: Option<AuthenticatorData>,
+struct ExternalBrowserChallengeResponse {
+    data: Option<ExternalBrowserChallenge>,
     message: Option<String>,
     success: bool,
 }
 
-fn authenticator_request_body(
-    username: &str,
-    account: &str,
-    redirect_port: u16,
-) -> serde_json::Value {
+fn challenge_request_body(username: &str, account: &str, redirect_port: u16) -> serde_json::Value {
     json!({
         "data": {
             "ACCOUNT_NAME": account,
@@ -212,185 +213,24 @@ fn open_auth_page(sso_url: &str, mode: BrowserLaunchMode) -> Result<()> {
     }
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum CallbackWaitError {
-    TimedOut,
-    ListenerStopped,
-}
-
 async fn resolve_payload_with_listener(
-    listener: RunningListener,
+    mut listener: RunningListener,
     timeout: Duration,
-) -> Result<CallbackPayload> {
-    let callback_result = wait_for_token(listener.payloads.clone(), timeout).await;
-    shutdown_listener(listener).await;
+) -> Result<BrowserCallbackPayload> {
+    let callback_result = listener.wait_for_payload(timeout).await;
+    listener.shutdown().await;
 
     match callback_result {
         Ok(payload) => Ok(payload),
         Err(CallbackWaitError::TimedOut) => {
             eprintln!("Callback was not received in time. Falling back to manual URL input.");
-            manual_token_flow().await
+            manual_redirect_input::manual_token_flow().await
         }
         Err(CallbackWaitError::ListenerStopped) => {
             eprintln!(
                 "Local callback listener stopped before receiving token. Continue with manual URL input."
             );
-            manual_token_flow().await
+            manual_redirect_input::manual_token_flow().await
         }
-    }
-}
-
-async fn wait_for_token(
-    rx: tokio::sync::watch::Receiver<Option<CallbackPayload>>,
-    timeout: Duration,
-) -> std::result::Result<CallbackPayload, CallbackWaitError> {
-    match time::timeout(timeout, wait_for_token_inner(rx)).await {
-        Ok(res) => res,
-        Err(_) => Err(CallbackWaitError::TimedOut),
-    }
-}
-
-async fn wait_for_token_inner(
-    rx: tokio::sync::watch::Receiver<Option<CallbackPayload>>,
-) -> std::result::Result<CallbackPayload, CallbackWaitError> {
-    let mut rx = rx.clone();
-    if let Some(payload) = rx.borrow().clone() {
-        return Ok(payload);
-    }
-
-    loop {
-        if rx.changed().await.is_err() {
-            return Err(CallbackWaitError::ListenerStopped);
-        }
-        if let Some(payload) = rx.borrow().clone() {
-            return Ok(payload);
-        }
-    }
-}
-
-async fn shutdown_listener(listener: RunningListener) {
-    let _ = listener.shutdown.send(());
-    let _ = listener.handle.await;
-}
-
-async fn manual_token_flow() -> Result<CallbackPayload> {
-    tokio::task::spawn_blocking(manual_token_flow_blocking)
-        .await
-        .map_err(|e| Error::Communication(format!("manual input task failed: {e}")))?
-}
-
-fn manual_token_flow_blocking() -> Result<CallbackPayload> {
-    eprintln!(
-        "After completing authentication, paste the URL you were redirected to (not logged)."
-    );
-
-    eprint!("Redirected URL: ");
-    let _ = io::stderr().flush();
-    let input = read_redirected_url_line()?;
-
-    eprintln!("Received redirected URL input. Continuing authentication...");
-
-    payload_from_redirect_input(input.trim())
-}
-
-fn read_redirected_url_line() -> Result<String> {
-    #[cfg(unix)]
-    if let Some(result) = manual_input_unix::try_read_redirected_url_line_noncanonical() {
-        return result;
-    }
-
-    // TODO(windows): investigate whether long redirected URLs can be truncated in
-    // canonical console input mode and add a Windows-specific raw/non-canonical fallback.
-
-    let mut input = String::new();
-    io::stdin()
-        .read_line(&mut input)
-        .map_err(|e| Error::Communication(format!("failed to read input: {e}")))?;
-
-    validate_redirected_url_input(input)
-}
-
-fn validate_redirected_url_input(input: String) -> Result<String> {
-    if input.trim().is_empty() {
-        return Err(Error::Communication(
-            "No redirected URL was provided".to_string(),
-        ));
-    }
-    Ok(input)
-}
-
-fn extract_payload_from_url(url: &str) -> Option<CallbackPayload> {
-    let parsed = Url::parse(url).ok()?;
-    let query = parse_token_and_consent_from_pairs(parsed.query_pairs());
-    let fragment = parsed
-        .fragment()
-        .map(|frag| {
-            parse_token_and_consent_from_pairs(url::form_urlencoded::parse(frag.as_bytes()))
-        })
-        .unwrap_or_default();
-
-    let token = query.token.or(fragment.token)?;
-    let consent = query.consent.or(fragment.consent);
-
-    Some(CallbackPayload { token, consent })
-}
-
-fn payload_from_redirect_input(input: &str) -> Result<CallbackPayload> {
-    extract_payload_from_url(input).ok_or_else(|| {
-        Error::Communication(
-            "Unable to extract token from redirected URL (expected query or fragment token=...)"
-                .to_string(),
-        )
-    })
-}
-
-#[cfg(test)]
-mod tests {
-    use super::payload_from_redirect_input;
-
-    #[test]
-    fn payload_from_redirect_input_extracts_token() {
-        let url = "https://example.test/callback?token=abc123&other=1";
-        let payload = payload_from_redirect_input(url).unwrap();
-        assert_eq!(payload.token, "abc123");
-        assert_eq!(payload.consent, None);
-    }
-
-    #[test]
-    fn payload_from_redirect_input_extracts_consent() {
-        let url = "https://example.test/callback?token=abc123&consent=false";
-        let payload = payload_from_redirect_input(url).unwrap();
-        assert_eq!(payload.token, "abc123");
-        assert_eq!(payload.consent, Some(false));
-    }
-
-    #[test]
-    fn payload_from_redirect_input_extracts_token_from_fragment() {
-        let url = "https://example.test/callback#token=abc123&consent=true";
-        let payload = payload_from_redirect_input(url).unwrap();
-        assert_eq!(payload.token, "abc123");
-        assert_eq!(payload.consent, Some(true));
-    }
-
-    #[test]
-    fn payload_from_redirect_input_prefers_query_token_over_fragment_token() {
-        let url = "https://example.test/callback?token=query_token#token=fragment_token";
-        let payload = payload_from_redirect_input(url).unwrap();
-        assert_eq!(payload.token, "query_token");
-    }
-
-    #[test]
-    fn payload_from_redirect_input_keeps_query_token_and_query_consent() {
-        let url = "https://example.test/callback?token=query_token&consent=false#consent=true";
-        let payload = payload_from_redirect_input(url).unwrap();
-        assert_eq!(payload.token, "query_token");
-        assert_eq!(payload.consent, Some(false));
-    }
-
-    #[test]
-    fn payload_from_redirect_input_fails_without_token() {
-        let url = "https://example.test/callback?missing=true";
-        let err = payload_from_redirect_input(url).unwrap_err();
-        assert!(format!("{err}").contains("Unable to extract token"));
     }
 }
