+++ src/external_browser_listener.rs
@@ -1,52 +1,31 @@
-use std::convert::Infallible;
-use std::error::Error;
-use std::net::{IpAddr, Ipv4Addr, SocketAddr};
-use std::sync::Arc;
+mod handler;
+mod runtime;
 
-use http_body_util::{BodyExt, Full};
-use hyper::body::{Bytes, Incoming};
-use hyper::header::{
-    ACCESS_CONTROL_REQUEST_HEADERS, CONTENT_LENGTH, CONTENT_TYPE, HeaderValue, ORIGIN, VARY,
+use std::{
+    error::Error,
+    net::{IpAddr, Ipv4Addr, SocketAddr},
+    time::Duration,
 };
-use hyper::http::StatusCode;
-use hyper::server::conn::http1;
-use hyper::service::service_fn;
-use hyper::{Method, Request, Response};
-use hyper_util::rt::TokioIo;
-use serde::Deserialize;
-use socket2::{Domain, Protocol, Socket, Type};
+
 use tokio::{
-    net::TcpListener,
-    sync::Mutex,
     sync::{oneshot, watch},
-    task::{JoinError, JoinHandle, JoinSet},
+    task::JoinHandle,
 };
 
-use crate::external_browser_payload::{ParsedTokenAndConsent, parse_token_and_consent_from_pairs};
+use crate::external_browser_payload::BrowserCallbackPayload;
+
+use runtime::{ListenerRuntime, bind_listener, build_origin};
 
 type ListenerError = Box<dyn Error + Send + Sync>;
-type ConnectionTaskResult = Result<(), ListenerError>;
 
-#[derive(Debug, Deserialize)]
-struct TokenPayload {
-    token: Option<String>,
-    consent: Option<bool>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct CallbackPayload {
-    pub token: String,
-    pub consent: Option<bool>,
-}
-
-pub struct ListenerConfig {
+pub(crate) struct ListenerConfig {
     /// Application name used in the human-facing HTML response.
     ///
     /// This corresponds to Python connector's `AuthByWebBrowser._application`.
-    pub application: Option<String>,
-    pub host: IpAddr,
-    pub port: u16,
-    pub protocol: String,
+    application: Option<String>,
+    host: IpAddr,
+    port: u16,
+    protocol: String,
 }
 
 impl Default for ListenerConfig {
@@ -60,222 +39,102 @@ impl Default for ListenerConfig {
     }
 }
 
-pub struct RunningListener {
-    pub addr: SocketAddr,
-    pub shutdown: oneshot::Sender<()>,
-    pub handle: JoinHandle<Result<(), ListenerError>>,
-    pub payloads: watch::Receiver<Option<CallbackPayload>>,
-}
-
-struct ListenerShared {
-    expected_origin: String,
-    application: Option<String>,
-    payload_tx: watch::Sender<Option<CallbackPayload>>,
-    validated_origin: Mutex<Option<String>>,
-}
-
-impl ListenerShared {
-    fn new(
-        expected_origin: String,
+impl ListenerConfig {
+    pub(crate) fn new(
         application: Option<String>,
-        payload_tx: watch::Sender<Option<CallbackPayload>>,
+        host: IpAddr,
+        port: u16,
+        protocol: String,
     ) -> Self {
         Self {
-            expected_origin,
             application,
-            payload_tx,
-            validated_origin: Mutex::new(None),
-        }
-    }
-
-    async fn handle(&self, req: Request<Incoming>) -> Result<Response<Full<Bytes>>, Infallible> {
-        let request_origin = header_to_string(req.headers().get(ORIGIN));
-
-        match *req.method() {
-            Method::OPTIONS => self.handle_preflight(request_origin, &req).await,
-            Method::GET => {
-                let payload = extract_payload_from_query(req.uri().query());
-                self.handle_callback(payload).await
-            }
-            Method::POST => {
-                let whole_body = req
-                    .into_body()
-                    .collect()
-                    .await
-                    .map(|c| c.to_bytes())
-                    .unwrap_or_default();
-                let payload = extract_payload_from_body(&whole_body);
-                self.handle_callback(payload).await
-            }
-            _ => Ok(Response::builder()
-                .status(StatusCode::METHOD_NOT_ALLOWED)
-                .body(Full::new(Bytes::from_static(b"method not allowed")))
-                .expect("building method not allowed response failed")),
-        }
-    }
-
-    async fn handle_preflight(
-        &self,
-        request_origin: Option<String>,
-        req: &Request<Incoming>,
-    ) -> Result<Response<Full<Bytes>>, Infallible> {
-        let expected_origin = self.expected_origin.clone();
-        let allowed_origin =
-            request_origin.filter(|origin| origin_allowed_value(origin, &expected_origin));
-        if allowed_origin.is_none() {
-            return Ok(forbidden());
-        }
-
-        self.set_validated_origin(allowed_origin.clone()).await;
-
-        let allow_headers =
-            requested_headers(req).unwrap_or_else(|| "Content-Type, Origin".to_string());
-        Ok(preflight_cors_response(
-            Response::builder().status(StatusCode::OK),
-            allowed_origin
-                .as_deref()
-                .unwrap_or(expected_origin.as_str()),
-            &allow_headers,
-        )
-        .header(VARY, "Accept-Encoding, Origin")
-        .header(CONTENT_LENGTH, "0")
-        .body(Full::new(Bytes::new()))
-        .expect("building preflight response failed"))
-    }
-
-    async fn handle_callback(
-        &self,
-        payload: Option<CallbackPayload>,
-    ) -> Result<Response<Full<Bytes>>, Infallible> {
-        self.publish_payload(&payload);
-        let stored_origin = self.validated_origin().await;
-        Ok(ok_callback_response(
-            payload.as_ref(),
-            stored_origin.as_deref(),
-            self.application.as_deref(),
-        ))
-    }
-
-    fn publish_payload(&self, payload: &Option<CallbackPayload>) {
-        if let Some(payload) = payload {
-            let _ = self.payload_tx.send(Some(payload.clone()));
-        }
-    }
-
-    async fn set_validated_origin(&self, origin: Option<String>) {
-        let mut guard = self.validated_origin.lock().await;
-        *guard = origin;
-    }
-
-    async fn validated_origin(&self) -> Option<String> {
-        self.validated_origin.lock().await.clone()
-    }
-}
-
-struct ConnectionTasks {
-    shutdown_tx: watch::Sender<bool>,
-    shutdown_rx: watch::Receiver<bool>,
-    tasks: JoinSet<ConnectionTaskResult>,
-}
-
-impl ConnectionTasks {
-    fn new() -> Self {
-        let (shutdown_tx, shutdown_rx) = watch::channel(false);
-        Self {
-            shutdown_tx,
-            shutdown_rx,
-            tasks: JoinSet::new(),
-        }
-    }
-
-    fn is_empty(&self) -> bool {
-        self.tasks.is_empty()
-    }
-
-    fn spawn(&mut self, stream: tokio::net::TcpStream, shared: Arc<ListenerShared>) {
-        let shutdown = self.shutdown_rx.clone();
-        self.tasks
-            .spawn(async move { run_connection(stream, shared, shutdown).await });
-    }
-
-    async fn join_next(&mut self) -> Option<Result<ConnectionTaskResult, JoinError>> {
-        self.tasks.join_next().await
-    }
-
-    async fn shutdown_and_drain(&mut self) {
-        let _ = self.shutdown_tx.send(true);
-        while !self.is_empty() {
-            log_connection_task_result(self.join_next().await);
+            host,
+            port,
+            protocol,
         }
     }
 }
 
-struct ListenerRuntime {
-    listener: TcpListener,
-    shared: Arc<ListenerShared>,
-    shutdown: oneshot::Receiver<()>,
-    connections: ConnectionTasks,
+pub(crate) struct RunningListener {
+    addr: SocketAddr,
+    shutdown: oneshot::Sender<()>,
+    handle: JoinHandle<Result<(), ListenerError>>,
+    payloads: watch::Receiver<Option<BrowserCallbackPayload>>,
 }
 
-impl ListenerRuntime {
-    fn new(
-        listener: TcpListener,
-        expected_origin: String,
-        application: Option<String>,
-        shutdown: oneshot::Receiver<()>,
-        payload_tx: watch::Sender<Option<CallbackPayload>>,
-    ) -> Self {
-        Self {
-            listener,
-            shared: Arc::new(ListenerShared::new(
-                expected_origin,
-                application,
-                payload_tx,
-            )),
-            shutdown,
-            connections: ConnectionTasks::new(),
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub(crate) enum CallbackWaitError {
+    TimedOut,
+    ListenerStopped,
+}
+
+impl RunningListener {
+    pub(crate) fn redirect_port(&self) -> u16 {
+        self.addr.port()
+    }
+
+    pub(crate) async fn wait_for_payload(
+        &mut self,
+        timeout: Duration,
+    ) -> std::result::Result<BrowserCallbackPayload, CallbackWaitError> {
+        match tokio::time::timeout(timeout, Self::wait_for_payload_inner(&mut self.payloads)).await
+        {
+            Ok(res) => res,
+            Err(_) => Err(CallbackWaitError::TimedOut),
         }
     }
 
-    async fn run(mut self) -> Result<(), ListenerError> {
+    async fn wait_for_payload_inner(
+        rx: &mut watch::Receiver<Option<BrowserCallbackPayload>>,
+    ) -> std::result::Result<BrowserCallbackPayload, CallbackWaitError> {
+        if let Some(payload) = rx.borrow().clone() {
+            return Ok(payload);
+        }
+
         loop {
-            tokio::select! {
-                join_result = self.connections.join_next(), if !self.connections.is_empty() => {
-                    log_connection_task_result(join_result);
-                }
-                res = self.listener.accept() => {
-                    let (stream, _peer) = match res {
-                        Ok(v) => v,
-                        Err(e) => {
-                            eprintln!("accept error: {e}");
-                            continue;
-                        }
-                    };
-                    self.connections.spawn(stream, self.shared.clone());
-                }
-                _ = &mut self.shutdown => {
-                    break;
-                }
+            if rx.changed().await.is_err() {
+                return Err(CallbackWaitError::ListenerStopped);
+            }
+            if let Some(payload) = rx.borrow().clone() {
+                return Ok(payload);
             }
         }
+    }
 
-        drop(self.listener);
-        self.connections.shutdown_and_drain().await;
-        Ok(())
+    pub(crate) async fn shutdown(self) {
+        let _ = self.shutdown.send(());
+        let _ = self.handle.await;
+    }
+
+    #[cfg(test)]
+    fn addr(&self) -> SocketAddr {
+        self.addr
+    }
+
+    #[cfg(test)]
+    #[allow(clippy::type_complexity)]
+    fn into_parts(
+        self,
+    ) -> (
+        SocketAddr,
+        oneshot::Sender<()>,
+        JoinHandle<Result<(), ListenerError>>,
+        watch::Receiver<Option<BrowserCallbackPayload>>,
+    ) {
+        (self.addr, self.shutdown, self.handle, self.payloads)
     }
 }
 
-pub async fn spawn_listener(cfg: ListenerConfig) -> Result<RunningListener, ListenerError> {
+pub(crate) async fn spawn_listener(cfg: ListenerConfig) -> Result<RunningListener, ListenerError> {
     let (listener, local_addr) = bind_listener(&cfg)?;
     let expected_origin = build_origin(&cfg.protocol, cfg.host, local_addr.port());
-    let application = cfg.application.clone();
     let (shutdown_tx, shutdown_rx) = oneshot::channel();
     let (payload_tx, payload_rx) = watch::channel(None);
     let handle = tokio::spawn(
         ListenerRuntime::new(
             listener,
             expected_origin,
-            application,
+            cfg.application,
             shutdown_rx,
             payload_tx,
         )
@@ -290,200 +149,18 @@ pub async fn spawn_listener(cfg: ListenerConfig) -> Result<RunningListener, List
     })
 }
 
-async fn run_connection(
-    stream: tokio::net::TcpStream,
-    shared: Arc<ListenerShared>,
-    mut shutdown: watch::Receiver<bool>,
-) -> ConnectionTaskResult {
-    let io = TokioIo::new(stream);
-    let svc = service_fn(move |req| {
-        let shared = shared.clone();
-        async move { shared.handle(req).await }
-    });
-    let conn = http1::Builder::new().serve_connection(io, svc);
-    tokio::pin!(conn);
-
-    tokio::select! {
-        res = conn.as_mut() => {
-            res.map_err(|err| Box::new(err) as ListenerError)
-        }
-        _ = shutdown.changed() => {
-            conn.as_mut().graceful_shutdown();
-            conn.await.map_err(|err| Box::new(err) as ListenerError)
-        }
-    }
-}
-
-fn log_connection_task_result(join_result: Option<Result<ConnectionTaskResult, JoinError>>) {
-    match join_result {
-        Some(Ok(Ok(()))) => {}
-        Some(Ok(Err(err))) => eprintln!("serve error: {err}"),
-        Some(Err(err)) => eprintln!("connection task join error: {err}"),
-        None => {}
-    }
-}
-
-fn extract_payload_from_body(body: &Bytes) -> Option<CallbackPayload> {
-    // 1) JSON: {"token": "...", "consent": true}
-    serde_json::from_slice::<TokenPayload>(body)
-        .ok()
-        .and_then(|parsed| {
-            parsed.token.map(|token| CallbackPayload {
-                token,
-                consent: parsed.consent,
-            })
-        })
-        // 2) x-www-form-urlencoded / key=value fallback
-        .or_else(|| {
-            std::str::from_utf8(body)
-                .ok()
-                .map(|body_str| {
-                    parse_token_and_consent_from_pairs(url::form_urlencoded::parse(
-                        body_str.as_bytes(),
-                    ))
-                })
-                .and_then(parsed_to_payload)
-        })
-}
-
-fn preflight_cors_response(
-    mut builder: hyper::http::response::Builder,
-    allowed_origin: &str,
-    allow_headers: &str,
-) -> hyper::http::response::Builder {
-    builder = builder
-        .header("Access-Control-Allow-Origin", allowed_origin)
-        .header("Access-Control-Allow-Methods", "POST, GET")
-        .header("Access-Control-Allow-Headers", allow_headers)
-        .header("Access-Control-Max-Age", "86400");
-    builder
-}
-
-fn ok_callback_response(
-    payload: Option<&CallbackPayload>,
-    stored_origin: Option<&str>,
-    application: Option<&str>,
-) -> Response<Full<Bytes>> {
-    if payload.is_none() {
-        return Response::builder()
-            .status(StatusCode::OK)
-            .header(CONTENT_TYPE, "text/plain")
-            .header(CONTENT_LENGTH, "17")
-            .body(Full::new(Bytes::from_static(b"no token provided")))
-            .expect("building no token provided response failed");
-    }
-
-    // - If a validated Origin exists (OPTIONS preflight succeeded), return JSON with consent only.
-    // - Otherwise, return a simple HTML page telling the user they can close the window.
-    let cors_mode = stored_origin.is_some();
-    let (content_type, body) = if cors_mode {
-        let consent = payload.and_then(|p| p.consent).unwrap_or(true);
-        let msg = serde_json::json!({"consent": consent}).to_string();
-        ("text/html", msg)
-    } else {
-        let app_line = application
-            .map(str::trim)
-            .filter(|s| !s.is_empty())
-            .map(|app| format!("Client application: {app}.\n"))
-            .unwrap_or_default();
-        let msg = format!(
-            r#"<!DOCTYPE html><html><head><meta charset=\"UTF-8\"/>
-<link rel=\"icon\" href=\"data:,\">
-<title>SAML Response for Snowflake</title></head>
-<body>
-Your identity was confirmed and propagated to Snowflake.
-{app_line}You can close this window now and go back where you started from.
-</body></html>"#
-        );
-        ("text/html", msg)
-    };
-
-    let mut builder = Response::builder()
-        .status(StatusCode::OK)
-        .header(CONTENT_TYPE, content_type)
-        .header(CONTENT_LENGTH, body.len().to_string());
-
-    if let Some(origin) = stored_origin {
-        builder = builder
-            .header("Access-Control-Allow-Origin", origin)
-            .header(VARY, "Accept-Encoding, Origin");
-    }
-
-    builder
-        .body(Full::new(Bytes::from(body)))
-        .expect("building OK callback response failed")
-}
-
-fn forbidden() -> Response<Full<Bytes>> {
-    Response::builder()
-        .status(StatusCode::FORBIDDEN)
-        .body(Full::new(Bytes::from_static(b"origin not allowed")))
-        .expect("building forbidden response failed")
-}
-
-fn extract_payload_from_query(query: Option<&str>) -> Option<CallbackPayload> {
-    query
-        .map(|q| parse_token_and_consent_from_pairs(url::form_urlencoded::parse(q.as_bytes())))
-        .and_then(parsed_to_payload)
-}
-
-fn parsed_to_payload(parsed: ParsedTokenAndConsent) -> Option<CallbackPayload> {
-    parsed.token.map(|token| CallbackPayload {
-        token,
-        consent: parsed.consent,
-    })
-}
-
-fn build_origin(protocol: &str, host: IpAddr, port: u16) -> String {
-    format!("{protocol}://{}", SocketAddr::new(host, port))
-}
-
-fn bind_listener(
-    cfg: &ListenerConfig,
-) -> Result<(TcpListener, SocketAddr), Box<dyn Error + Send + Sync>> {
-    let target_addr = SocketAddr::new(cfg.host, cfg.port);
-
-    let domain = match target_addr {
-        SocketAddr::V4(_) => Domain::IPV4,
-        SocketAddr::V6(_) => Domain::IPV6,
-    };
-
-    let socket = Socket::new(domain, Type::STREAM, Some(Protocol::TCP))?;
-    socket.set_nonblocking(true)?;
-    socket.set_reuse_address(true)?;
-    socket.bind(&target_addr.into())?;
-    socket.listen(128)?;
-
-    let listener = TcpListener::from_std(socket.into())?;
-    let local_addr = listener.local_addr()?;
-    Ok((listener, local_addr))
-}
-
-fn header_to_string(value: Option<&HeaderValue>) -> Option<String> {
-    value.and_then(|v| v.to_str().ok()).map(|s| s.to_string())
-}
-
-fn origin_allowed_value(origin: &str, expected: &str) -> bool {
-    origin.eq_ignore_ascii_case(expected)
-}
-
-fn requested_headers(req: &Request<Incoming>) -> Option<String> {
-    req.headers()
-        .get(ACCESS_CONTROL_REQUEST_HEADERS)
-        .and_then(|h| h.to_str().ok())
-        .map(|s| s.to_string())
-}
-
 #[cfg(test)]
 mod tests {
-    use std::io;
-    use std::net::TcpListener as StdTcpListener;
+    use std::{io, net::TcpListener as StdTcpListener};
 
     use reqwest::StatusCode as ReqwestStatusCode;
-    use tokio::io::{AsyncReadExt, AsyncWriteExt};
-    use tokio::net::TcpStream;
-    use tokio::time::{Duration, sleep};
+    use tokio::{
+        io::{AsyncReadExt, AsyncWriteExt},
+        net::TcpStream,
+        time::{Duration, sleep},
+    };
 
+    use super::runtime::bind_listener;
     use super::*;
 
     async fn read_http_response(stream: &mut TcpStream) -> io::Result<String> {
@@ -550,101 +227,16 @@ mod tests {
         })
         .await
         .unwrap();
-        let base = format!("http://{}", running.addr);
+        let (addr, shutdown, handle, _) = running.into_parts();
+        let base = format!("http://{addr}");
 
         // give the server a moment to start accepting
         sleep(Duration::from_millis(50)).await;
 
         test(base).await;
 
-        let _ = running.shutdown.send(());
-        let _ = running.handle.await;
-    }
-
-    #[test]
-    fn extract_payload_from_body_json() {
-        let body = Bytes::from_static(br#"{"token":"t","consent":false}"#);
-        let payload = extract_payload_from_body(&body).unwrap();
-        assert_eq!(payload.token, "t");
-        assert_eq!(payload.consent, Some(false));
-    }
-
-    #[test]
-    fn extract_payload_from_body_form_urlencoded() {
-        let body = Bytes::from_static(b"token=t&consent=true");
-        let payload = extract_payload_from_body(&body).unwrap();
-        assert_eq!(payload.token, "t");
-        assert_eq!(payload.consent, Some(true));
-    }
-
-    #[test]
-    fn extract_payload_from_query_parses_consent() {
-        let payload = extract_payload_from_query(Some("token=t&consent=false")).unwrap();
-        assert_eq!(payload.token, "t");
-        assert_eq!(payload.consent, Some(false));
-    }
-
-    #[test]
-    fn extract_payload_from_query_prefers_first_token() {
-        let payload = extract_payload_from_query(Some("token=first&token=second")).unwrap();
-        assert_eq!(payload.token, "first");
-    }
-
-    #[tokio::test]
-    async fn ok_callback_response_cors_mode_returns_consent_only() {
-        let payload = CallbackPayload {
-            token: "secret".to_string(),
-            consent: Some(false),
-        };
-        let resp = ok_callback_response(Some(&payload), Some("http://localhost:1"), None);
-        assert_eq!(resp.status(), StatusCode::OK);
-        let headers = resp.headers().clone();
-
-        assert_eq!(
-            headers.get(CONTENT_TYPE).unwrap().to_str().unwrap(),
-            "text/html"
-        );
-        assert_eq!(
-            headers
-                .get("access-control-allow-origin")
-                .unwrap()
-                .to_str()
-                .unwrap(),
-            "http://localhost:1"
-        );
-        assert_eq!(
-            headers.get(VARY).unwrap().to_str().unwrap(),
-            "Accept-Encoding, Origin"
-        );
-
-        let body = resp.into_body().collect().await.unwrap().to_bytes();
-        let body_str = std::str::from_utf8(&body).unwrap();
-        assert!(body_str.contains("\"consent\":false"));
-        assert!(!body_str.contains("secret"));
-
-        let expected_len = body.len().to_string();
-        assert_eq!(
-            headers.get(CONTENT_LENGTH).unwrap().to_str().unwrap(),
-            expected_len
-        );
-    }
-
-    #[tokio::test]
-    async fn ok_callback_response_html_includes_app_line() {
-        let payload = CallbackPayload {
-            token: "t".to_string(),
-            consent: None,
-        };
-        let resp = ok_callback_response(Some(&payload), None, Some("myapp"));
-        assert_eq!(resp.status(), StatusCode::OK);
-        let headers = resp.headers().clone();
-        assert!(headers.get("access-control-allow-origin").is_none());
-        assert!(headers.get(VARY).is_none());
-        let body = resp.into_body().collect().await.unwrap().to_bytes();
-        let body_str = std::str::from_utf8(&body).unwrap();
-        assert!(body_str.contains("You can close this window now"));
-        assert!(body_str.contains("Client application: myapp"));
-        assert!(!body_str.contains("token"));
+        let _ = shutdown.send(());
+        let _ = handle.await;
     }
 
     #[tokio::test]
@@ -690,7 +282,7 @@ mod tests {
                     .unwrap()
                     .to_str()
                     .unwrap(),
-                "text/html"
+                "application/json"
             );
             let body = resp.text().await.unwrap();
             assert!(body.contains("\"consent\""));
@@ -733,6 +325,7 @@ mod tests {
                 .await
                 .unwrap();
             assert_eq!(resp.status(), ReqwestStatusCode::OK);
+
             let body = resp.text().await.unwrap();
             assert!(body.contains("You can close this window now"));
             assert!(body.contains("Client application: testapp"));
@@ -767,6 +360,7 @@ mod tests {
                 .await
                 .unwrap();
             assert_eq!(resp.status(), ReqwestStatusCode::OK);
+
             let body = resp.text().await.unwrap();
             assert!(body.contains("You can close this window now"));
             assert!(body.contains("Client application: testapp"));
@@ -788,6 +382,7 @@ mod tests {
                 .await
                 .unwrap();
             assert_eq!(resp.status(), ReqwestStatusCode::OK);
+
             let headers = resp.headers();
             assert_eq!(
                 headers.get("vary").unwrap().to_str().unwrap(),
@@ -816,8 +411,9 @@ mod tests {
     #[tokio::test]
     async fn post_json_propagates_consent_in_payload() {
         let running = spawn_listener(ListenerConfig::default()).await.unwrap();
-        let base = format!("http://{}", running.addr);
-        let mut rx = running.payloads.clone();
+        let (addr, shutdown, handle, payloads) = running.into_parts();
+        let base = format!("http://{addr}");
+        let mut rx = payloads;
 
         let client = reqwest::Client::new();
         let resp = client
@@ -833,15 +429,16 @@ mod tests {
         assert_eq!(payload.token, "example");
         assert_eq!(payload.consent, Some(false));
 
-        let _ = running.shutdown.send(());
-        let _ = running.handle.await;
+        let _ = shutdown.send(());
+        let _ = handle.await;
     }
 
     #[tokio::test]
     async fn form_encoded_payload_is_extracted() {
         let running = spawn_listener(ListenerConfig::default()).await.unwrap();
-        let base = format!("http://{}", running.addr);
-        let mut rx = running.payloads.clone();
+        let (addr, shutdown, handle, payloads) = running.into_parts();
+        let base = format!("http://{addr}");
+        let mut rx = payloads;
 
         let client = reqwest::Client::new();
         let resp = client
@@ -858,15 +455,16 @@ mod tests {
         assert_eq!(payload.token, "example");
         assert_eq!(payload.consent, Some(true));
 
-        let _ = running.shutdown.send(());
-        let _ = running.handle.await;
+        let _ = shutdown.send(());
+        let _ = handle.await;
     }
 
     #[tokio::test]
     async fn query_payload_is_extracted() {
         let running = spawn_listener(ListenerConfig::default()).await.unwrap();
-        let base = format!("http://{}", running.addr);
-        let mut rx = running.payloads.clone();
+        let (addr, shutdown, handle, payloads) = running.into_parts();
+        let base = format!("http://{addr}");
+        let mut rx = payloads;
 
         let client = reqwest::Client::new();
         let resp = client
@@ -881,15 +479,15 @@ mod tests {
         assert_eq!(payload.token, "example");
         assert_eq!(payload.consent, Some(false));
 
-        let _ = running.shutdown.send(());
-        let _ = running.handle.await;
+        let _ = shutdown.send(());
+        let _ = handle.await;
     }
 
     #[tokio::test]
     async fn dropping_running_listener_releases_bound_port() {
         let running = spawn_listener(ListenerConfig::default()).await.unwrap();
-        let port = running.addr.port();
-        let host = running.addr.ip();
+        let port = running.redirect_port();
+        let host = running.addr().ip();
 
         drop(running);
 
@@ -932,11 +530,11 @@ mod tests {
         .await
         .unwrap();
 
-        let mut stale_rx = running1.payloads.clone();
-        let mut stream = TcpStream::connect(running1.addr).await.unwrap();
+        let (addr1, shutdown1, handle1, payloads1) = running1.into_parts();
+        let mut stale_rx = payloads1;
+        let mut stream = TcpStream::connect(addr1).await.unwrap();
         let first_request = format!(
-            "GET /callback?token=first HTTP/1.1\r\nHost: {}\r\nConnection: keep-alive\r\n\r\n",
-            running1.addr
+            "GET /callback?token=first HTTP/1.1\r\nHost: {addr1}\r\nConnection: keep-alive\r\n\r\n",
         );
         stream.write_all(first_request.as_bytes()).await.unwrap();
         let first_response = read_http_response(&mut stream).await.unwrap();
@@ -945,8 +543,8 @@ mod tests {
         let _ = stale_rx.changed().await;
         assert_eq!(stale_rx.borrow().as_ref().unwrap().token, "first");
 
-        let _ = running1.shutdown.send(());
-        let _ = running1.handle.await;
+        let _ = shutdown1.send(());
+        let _ = handle1.await;
 
         let mut closed_probe = [0_u8; 1];
         let close_result =
@@ -957,12 +555,12 @@ mod tests {
         assert_eq!(close_result, 0, "stale callback connection remained open");
 
         let running2 = spawn_listener(cfg).await.unwrap();
-        let mut next_rx = running2.payloads.clone();
+        let (addr2, shutdown2, handle2, payloads2) = running2.into_parts();
+        let mut next_rx = payloads2;
 
-        let mut second_stream = TcpStream::connect(running2.addr).await.unwrap();
+        let mut second_stream = TcpStream::connect(addr2).await.unwrap();
         let second_request = format!(
-            "GET /callback?token=second HTTP/1.1\r\nHost: {}\r\nConnection: keep-alive\r\n\r\n",
-            running2.addr
+            "GET /callback?token=second HTTP/1.1\r\nHost: {addr2}\r\nConnection: keep-alive\r\n\r\n",
         );
         second_stream
             .write_all(second_request.as_bytes())
@@ -979,7 +577,7 @@ mod tests {
         );
         assert_eq!(next_rx.borrow().as_ref().unwrap().token, "second");
 
-        let _ = running2.shutdown.send(());
-        let _ = running2.handle.await;
+        let _ = shutdown2.send(());
+        let _ = handle2.await;
     }
 }
